first try implementing multiple wallets
