+++ app/build.gradle
@@ -64,4 +64,7 @@ dependencies {
     // sqlcipher
     implementation "net.zetetic:android-database-sqlcipher:4.5.3"
     implementation "androidx.sqlite:sqlite:2.1.0"
+
+    // home fragment tabs (Wallets)
+    implementation "androidx.viewpager2:viewpager2:1.0.0"
 }
